Added authentication in the listing controller

+++ backend/controllers/listing-controller.js
@@ -2,7 +2,14 @@ const express = require("express");
 const router = express.Router();
 const { User } = require("../models");
 const { Listing } = require("../models");
+const Passage = require("@passageidentity/passage-node");
 require("dotenv").config();
+
+const passage = new Passage({
+    appID: process.env.PASSAGE_APP_ID,
+    apiKey: process.env.PASSAGE_API_KEY,
+    authStrategy: "HEADER",
+});
 
 // retrieve information from all listings
 router.get("/", async (req, res) => {
@@ -38,7 +45,6 @@ router.get("/", async (req, res) => {
         ...(category && {category: category}),
         ...(subtypes && {subtypes: subtypes}),
         ...(price && {price: price})
-
     };
 
     console.log(filter);
@@ -74,17 +80,32 @@ router.get("/user/:id", async (req, res) => {
 // creates user listing
 router.post("/", async (req, res) => {
     try {
-        const location = req.body.location
-        fetch(`https://www.mapquestapi.com/geocoding/v1/address?key=${process.env.GEOCODE_API_KEY}&location=${location}`)
-        .then((res) => res.json())
-        .then( async (json) => {
-            const zipInfo = (json.results[0].locations[0].latLng);
-            req.body.zipCoords = { type: "Point", coordinates: [zipInfo.lng, zipInfo.lat]};
-            const newListing = await Listing.create(req.body);
-            res.status(200).json(newListing);
-        })
-        .catch(console.error);
-    } catch (err) {
+        //////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+        const userID = await passage.authenticateRequest(req);
+        if (userID) {
+            // user is authenticated
+            const { email, phone, user_metadata } = await passage.user.get(userID);
+            const identifier = email ? email : phone;
+            const getUser = await User.findOne({passage_id: identifier});
+            if (getUser) { 
+                const location = req.body.location
+                fetch(`https://www.mapquestapi.com/geocoding/v1/address?key=${process.env.GEOCODE_API_KEY}&location=${location}`)
+                .then((res) => res.json())
+                .then( async (json) => {
+                    const zipInfo = (json.results[0].locations[0].latLng);
+                    req.body.zipCoords = { type: "Point", coordinates: [zipInfo.lng, zipInfo.lat]};
+                    req.body.userID = getUser._id;
+                    const newListing = await Listing.create(req.body);
+                    res.status(200).json(newListing);
+                }).catch((err) => {
+                    res.status(400).json({ error: err.message });
+                });
+            } else {
+                // logic to get meta data from passage and create listing
+                res.status(400).json({ error: "The user is not logged in to post things" });
+            };
+        }
+    }catch (err) {
         res.status(400).json({ error: err.message });
     }
 });
@@ -92,22 +113,50 @@ router.post("/", async (req, res) => {
 //edit Listing for user
 router.put("/:id", async (req, res) => {
   try {
-    const updatedListing = await Listing.findByIdAndUpdate(
-      req.params.id,
-      req.body,
-      { new: true, runValidators: true }
-    );
-    res.status(201).json(updatedListing);
+    const userID = await passage.authenticateRequest(req);
+    if (userID) {
+        const { email, phone, user_metadata } = await passage.user.get(userID);
+        const identifier = email ? email : phone;
+        const getUser = await User.findOne({passage_id: identifier});
+        if (getUser) {
+        if (req.body.location){
+            await fetch(`https://www.mapquestapi.com/geocoding/v1/address?key=${process.env.GEOCODE_API_KEY}&location=${req.body.location}`)
+            .then((res) => res.json())
+            .then(async (json) => {
+                const zipInfo = (json.results[0].locations[0].latLng);
+                req.body.zipCoords = { type: "Point", coordinates: [zipInfo.lng, zipInfo.lat] }
+            })
+        };
+        const updatedListing = await Listing.findOneAndUpdate({
+                _id: req.params.id,
+                userID: getUser._id
+            },
+            req.body,
+            { new: true, runValidators: true }
+        );
+        res.status(201).json(updatedListing);
+        }} else {
+        res.status(400).json({ error: "The user is not logged in to post things" });
+    }
   } catch (err) {
     res.status(400).json({ error: err });
   }
 });
 
-//delete Listing for usser
+//delete Listing for user
 router.delete("/:id", async (req, res) => {
   try {
-    const deletedListing = await Listing.findByIdAndDelete(req.params.id);
-    res.status(200).json(deletedListing);
+    const userID = await passage.authenticateRequest(req);
+    if (userID) {
+        // user is authenticated
+        const { email, phone, user_metadata } = await passage.user.get(userID);
+        const identifier = email ? email : phone;
+        const getUser = await User.findOne({passage_id: identifier});
+        const deletedListing = await Listing.findByIdAndDelete(req.params.id);
+        res.status(200).json(deletedListing);
+    }
+    // const deletedListing = await Listing.findByIdAndDelete(req.params.id);
+    // res.status(200).json(deletedListing);
   } catch (err) {
     res.status(400).json({ error: err });
   }
